requirements.txt
discord.py
starlette
uvicorn
python-dotenv
pynacl